+++ RELEASE-NOTES.md
@@ -41,6 +41,7 @@ END TEMPLATE-->
 
 * Made a new `IMetricsManager` interface with an `UpdateMetrics` event that can be used to update Prometheus metrics whenever they are scraped.
   * Also added a `metrics.update_interval` CVar to go along with this, when metrics are scraped without usage of Prometheus directly.
+* IoC now contains an `IMeterFactory` implementation that you can use to instantiate metric meters.
 
 ### Bugfixes